requirements.txt
aiohttp @ file:///opt/concourse/worker/volumes/live/c3cb6692-82cf-4cbe-4c9c-30944b7532e9/volume/aiohttp_1614360972016/work
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
boto==2.49.0
boto3 @ file:///tmp/build/80754af9/boto3_1614015198636/work
botocore @ file:///tmp/build/80754af9/botocore_1614361462893/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1611600262290/work
certifi==2020.12.5
cffi @ file:///opt/concourse/worker/volumes/live/0ef369cc-6ba0-47e7-75da-208c6400381d/volume/cffi_1613246948181/work
chardet @ file:///opt/concourse/worker/volumes/live/a5a7c56b-23cb-471d-7404-f4130b8aed33/volume/chardet_1605303184529/work
click==7.1.2
cryptography @ file:///opt/concourse/worker/volumes/live/61d16c80-891a-48c5-6376-6899ef51ea28/volume/cryptography_1613401441154/work
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1593129968237/work
djangorestframework==3.12.2
gensim @ file:///opt/concourse/worker/volumes/live/04a22e9f-8f9a-4d9c-50e4-f202b167c179/volume/gensim_1613994444097/work
google-api-core @ file:///C:/ci/google-api-core-split_1613980333946/work
google-auth @ file:///tmp/build/80754af9/google-auth_1613588361630/work
google-cloud-core @ file:///tmp/build/80754af9/google-cloud-core_1613414810663/work
google-cloud-storage @ file:///tmp/build/80754af9/google-cloud-storage_1614362409953/work
google-crc32c @ file:///opt/concourse/worker/volumes/live/423965c9-74d0-49ca-5750-745a67e15b38/volume/google-crc32c_1612242918626/work
google-resumable-media @ file:///tmp/build/80754af9/google-resumable-media_1608056254223/work
googleapis-common-protos @ file:///opt/concourse/worker/volumes/live/672c7a02-d061-4d6b-693d-bc534a88bbf1/volume/googleapis-common-protos-feedstock_1613725139744/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1612898609989/work
jsonfield==3.1.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///opt/concourse/worker/volumes/live/13289f20-f4bb-4e3d-7e57-60a41fceffd4/volume/multidict_1607367786469/work
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
numpy @ file:///opt/concourse/worker/volumes/live/5572694e-967a-4c0c-52cf-b53d43e72de9/volume/numpy_and_numpy_base_1603491881791/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.1.48
pdf2image @ file:///home/conda/feedstock_root/build_artifacts/pdf2image_1588278609827/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1593693221521/work
protobuf==3.14.0
psycopg2 @ file:///opt/concourse/worker/volumes/live/d12e789f-5b6c-4947-7cc2-187f3a9c9696/volume/psycopg2_1612298599991/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
pytesseract @ file:///home/conda/feedstock_root/build_artifacts/pytesseract_1605866581960/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
regex @ file:///opt/concourse/worker/volumes/live/7f106f75-0e11-45be-4c20-6b071e37c646/volume/regex_1602786678165/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
s3transfer @ file:///tmp/build/80754af9/s3transfer_1610404175646/work
scikit-learn==0.22.1
scipy==1.4.1
six @ file:///opt/concourse/worker/volumes/live/5b31cb27-1e37-4ca5-6e9f-86246eb206d2/volume/six_1605205320872/work
smart-open @ file:///tmp/build/80754af9/smart_open_1614023135845/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
Werkzeug==1.0.1
yarl @ file:///opt/concourse/worker/volumes/live/1ca49bfe-d6bc-4f62-4ca7-30c33d4f9b33/volume/yarl_1598045187281/work
